+++ src/main/java/web/SpringBoot2/dao/UsersDao.java
@@ -1,17 +1,14 @@
 package web.SpringBoot2.dao;
 
-
-
 import web.SpringBoot2.model.User;
-
 import java.util.List;
 
 public interface UsersDao {
      List<User> index();
-     User show(int id);
+     User show(Long id);
      void save(User user);
      void update(User updatedUsers);
-     void delete(int id);
+     void delete(Long id);
